Move Decode(options) from CRFSharpConsole to CRFSharpWrapper.csproj.
