+++ Core/CRFSharpWrapper/CRFSharpWrapper.csproj
@@ -66,6 +66,7 @@
   <ItemGroup>
     <Compile Include="Args.cs" />
     <Compile Include="CRFSharpHelper.cs" />
+    <Compile Include="DecoderDriver.cs" />
     <Compile Include="Decoder.cs" />
     <Compile Include="Encoder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
